apply port adapter pattern to mixpanel analytics
