+++ libs/adapters/src/index.ts
@@ -1,5 +1,6 @@
 export * from './express';
 export * from './hot-shots';
+export * from './mixpanel';
 export * from './rabbitmq';
 export * from './redis';
 export * from './typescript-logging';
