--- packages/commonwealth/server/controllers/server_analytics_methods/track.ts
@@ -1,8 +1,0 @@
-import { serverAnalyticsTrack } from '../../../shared/analytics/server-track';
-import { AnalyticsPayload } from '../../../shared/analytics/types';
-
-export type TrackOptions = Record<string, any> & AnalyticsPayload;
-
-export async function __track(options: TrackOptions) {
-  return serverAnalyticsTrack(options);
-}
