+++ packages/commonwealth/server/controllers/server_polls_methods/update_poll_vote.ts
@@ -8,7 +8,7 @@ import {
 } from '@hicommonwealth/model';
 import { MixpanelCommunityInteractionEvent } from '../../../shared/analytics/types';
 import { validateTopicGroupsMembership } from '../../util/requirementsModule/validateTopicGroupsMembership';
-import { TrackOptions } from '../server_analytics_methods/track';
+import { TrackOptions } from '../server_analytics_controller';
 import { ServerPollsController } from '../server_polls_controller';
 
 export const Errors = {
