+++ packages/commonwealth/server/controllers/server_topics_methods/create_topic.ts
@@ -7,7 +7,7 @@ import {
 import { sanitizeQuillText } from 'server/util/sanitizeQuillText';
 import { MixpanelCommunityInteractionEvent } from '../../../shared/analytics/types';
 import { validateOwner } from '../../util/validateOwner';
-import { TrackOptions } from '../server_analytics_methods/track';
+import { TrackOptions } from '../server_analytics_controller';
 import { ServerTopicsController } from '../server_topics_controller';
 
 export const Errors = {
